move filter serialization tests to kernel tests
