+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -86,25 +86,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "filter_dataset_serialization_test",
-    size = "medium",
-    srcs = ["filter_dataset_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python:math_ops",
-        "//tensorflow/python:sparse_tensor",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "//third_party/py/numpy",
-    ],
-)
-
-tf_py_test(
     name = "fixed_length_record_dataset_serialization_test",
     size = "medium",
     srcs = ["fixed_length_record_dataset_serialization_test.py"],
